requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pandas==2.2.2
tldextract==5.1.2
datasets==2.21.0
pyarrow==17.0.0
llama-cpp-python==0.3.7
